(LEARNVM-636) Disable mcollective
There's no direct way to disable mcollective. It seems like the most
plausible alternative is setting skip_tags in puppet.conf to skip all
resources tagged with puppet_enterprise::mcollective.

The change in this repository adds a parameter to set this option. To
use it, corresponding change will need to be made in the pltraining-bootstrap
repository to set the disable_mco parameter to true.